refactor : 예외처리 코드 정리
