+++ src/main/java/com/ybe/tr1ll1on/domain/accommodation/dto/request/AccommodationRequest.java
@@ -9,7 +9,6 @@ import java.time.LocalDate;
 
 @Getter
 @Setter
-@AllArgsConstructor
 @Builder
 public class AccommodationRequest {
     @NotNull
