+++ src/main/java/com/ybe/tr1ll1on/domain/cart/exception/CartException.java
@@ -3,8 +3,9 @@ package com.ybe.tr1ll1on.domain.cart.exception;
 import com.ybe.tr1ll1on.global.exception.ExceptionCode;
 import com.ybe.tr1ll1on.global.exception.TrillionException;
 
-public class UserNotFoundException extends TrillionException {
-    public UserNotFoundException(ExceptionCode errorCode) {
+public class CartException extends TrillionException {
+
+    public CartException(ExceptionCode errorCode) {
         super(errorCode);
     }
 }
