+++ src/main/java/com/ybe/tr1ll1on/domain/cart/exception/CartExceptionCode.java
@@ -7,12 +7,13 @@ import org.springframework.http.HttpStatus;
 
 @AllArgsConstructor
 @Getter
-public enum ProductNotExsitExceptionCode implements ExceptionCode {
+public enum CartExceptionCode implements ExceptionCode {
 
-    PRODUCT_NOT_FOUND(HttpStatus.NOT_FOUND, "PRODUCT_NOT_FOUND", "상품을 찾을 수 없습니다");
+    CARTID_NOT_FOUND(HttpStatus.NOT_FOUND, "CARTID_NOT_FOUND", "장바구니ID를 찾을 수 없습니다"),
+    CART_ITEM_ID_NOT_FOUND(HttpStatus.NOT_FOUND, "CART_ITEM_ID_NOT_FOUND", "장바구니상품ID를 찾을 수 없습니다")
+    ;
 
     private final HttpStatus status;
     private final String code;
     private final String msg;
-
 }
