--- src/main/java/com/ybe/tr1ll1on/domain/cart/exception/UserNotFoundExceptionCode.java
@@ -1,15 +1,0 @@
-package com.ybe.tr1ll1on.domain.cart.exception;
-
-import com.ybe.tr1ll1on.global.exception.ExceptionCode;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum UserNotFoundExceptionCode implements ExceptionCode {
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER_NOT_FOUND", "사용자를 찾을 수 없습니다");
-    private final HttpStatus status;
-    private final String code;
-    private final String msg;
-}
